Remove aui prefixes from property names
The prefix should not be used in property names. It is needed for
implementation in some metadata language; that is addressed in part by
#50 so this depends ion that pull request.

Examples are not touched and retain the prefix, since they are attribute
examples which would need the prefix. In the future we should provide
examples using multiple metatdata languages to help distinguish property
names from implementation of property names in metadata languages.

This closes #48.